Changed fav_icon to load from BraveCore sync instead and only if it fails to load, then fallback to iOS FavIconFetcher.
